fixed index visual bug and save bug
* indices are no longer calculated using row number
  * switching tabs while scrolled down no longer messes up 'Index' column (previously 'Id' column)
  * removed lag when switching tabs
* saving file with different name no longer deletes the original file

+++ SFVAnimationsEditor/ViewModel/MainViewModel.cs
@@ -197,8 +197,14 @@ namespace SFVAnimationsEditor.ViewModel
                 Console.WriteLine($"\n({DateTime.Now}) Save complete.");
 
                 // delete temp file
-                br.Dispose(); // release resources used by reader so the temp file can be deleted
-                File.Delete(tempName);
+                br.Dispose(); // release resources used by reader so the temp file can be deleted/renamed
+                if (FilePath != saveDialog.FileName)
+                    File.Move(tempName, FilePath); // rename tmp back to original if it wasn't overwritten
+                else
+                {
+                    File.Delete(tempName); // delete tmp if original was overwritten
+                    FilePath = saveDialog.FileName;
+                }
             }
             catch (Exception)
             {
